Notifications API (First version).
EmailNotification class uses System.Net to send email notifications.
The SendNotification method receives a list of strings with the receiver's email addresses, a subject string, a plain body string and a HTML body string to send as alternate view.
Can be improved by receiving a list of usernames instead of the email address, obtain the email address and name for each user and use that to send the notification.
Could also become an async method.