Fixed bug in delete button.
There are some cases where the id of one delete button is repeated between buttons. This makes that when a button is pressed, the selected option is not deleted but other.
Changed a little how the parent node of the button is found so there is no need to use IDs in buttons.